refactor: remove tail process in favor of exporter logging
Exposing the nginx access log via tail sometimes did not work, though we
have no clue about the concrete reason. So we dump it in favor of the
possibility to print the access log via the exporter which also tails
the log (and which seems to work in every case).